win32: drop support for floating subwindows
Floating subwindows, a.k.a. snapped-out windows are impossibly buggy UI
on Windows GTK as of 3.18 thru 3.22. There are hideous offsets between
cursor input and click target, and other deep, user-hostile interface
glitches. They work fine in Xorg, but on Windows we can't release in
this state.

Closes mypaint/mypaint#679.

[Cherry-pick of c566e99be1823128783359220337e54d9f89b32a from v1.2.x]

It would be good to revert this on master when we can figure out how.