workspace: add sidebar swapper button
Add a button to swap the current tool panel's sidebar side.  This is to
compensate for the reduced DnD ease and functionality on Windows since
127d3456d6b07ac38ae55729bdef80ce8a27293c, but it's useful for everyone.

Update the common buttons' tooltips and icons.

Also trying to get these buttons as small as possible, but things are
nevertheless looking a little cluttered in sidebars now.

Closes mypaint/mypaint#139