Add basic distutils support
We need to get rid of SCons. It's a mess.

https://wiki.debian.org/UpstreamGuide#SCons
https://wiki.gentoo.org/wiki/SCons

Let's use good old-fashioned distutils instead.
We might as well use the standard Python tool for the job,
and now that we're no longer building brushlib/libmypaint as a submodule,
it makes sense to switch to it.

Current state: the following just about works in MSYS2 "native" shells and
Linux:

    python2 setup.py build
    python2 setup.py install --prefix=/tmp/somewhere

you should then be able to run bin/mypaint[.py] after cd'ing there.
One major difference between this and SCons is that all the Python stuff
including the _mypaintlib.so extension now goes in $PREFIX/lib/mypaint. It
works fine though...