setup.py demo: use a throwaway install
Make the demo command install to a temporary tree which is deleted after
use. This helps keep the build area pristine.

The config area is a throwaway too, since we often need people to test
like that, and to reinforce the idea that this is a demo, not a
permanent way of running MyPaint.