Add Flatpak packaging
Flatpak is a great way to distribute applications in modern GNU/Linux
distributions.

To build a flatpak package locally, follow this steps:

    cd flatpak
    flatpak remote-add --no-gpg-verify --user my-repo ~/flatpak-repo
    flatpak-builder --repo=~/flatpak-repo --force-clean my-build org.mypaint.MyPaint.json
    flatpak --user install my-repo org.mypaint.MyPaint

Then it can be run from the command line like this:

    flatpak run org.mypaint.MyPaint

A flatpak repository can be created to easily install MyPaint from a
browser with one click. Is also possible to build a bundle for sharing
with `flatpak build-bundle`.  A central place for building and hosting
Flatpak builds named Flathub is in the works. And Alexander Larsson
showed interest in adding MyPaint to it.