externalapp: remove OpenWithDialog and use GTK+
Important note: this bumps the GTK+ dependency from 3.12 to 3.22 .

Since GTK+ 3.22 there is a `Gtk.show_uri_on_window()` that does the job
implemented in OpenWithDialog.

This not only removes a lot of custom code.  Is also a must to run
MyPaint in sandboxed contexts,like Flatpak or Snappy.  Because under the
hood, `Gtk.show_uri_on_window()` will automatically use the `open_uri`
portal from xdg-desktop-portal in those situations.