Add API to allow Cairo drawing on layers.
This commit adds new "cairo_request" context manager methods that allow
MyPaint surfaces and painting layers to be altered using a Cairo drawing
context. It is now possible to use the Cairo API to:

1. Draw directly into an 8-bpc RGBA pixbuf-surface.

2. Draw over a 15-bpc RGBA tiled surface or a painting layer,
   compositing the result back into it with one of
   MyPaint's layer modes (isolated rendering, basically).

3. Draw directly into a tiled surface or a painting layer,
   which lets Cairo's operators chew on the pixel data in
   8-bpc RGBA space and mess with its alpha.

The API operates as a context manager because it has to use several
temporary objects for format conversions.

Closes mypaint/mypaint#78 (no strokemap interaction though, and no speed
guarantees).

Helpful for mypaint/mypaint#150 and mypaint/mypaint#31?