Make blit_tile_into use the surface's impl.
The blit/copy method of surface-backed layers was needlessly using a
compositing pathway. This is something of an unused code path currently,
but I have an upcoming use for layer's blit_tile_into().

This looks legit, but I'm putting this into its own commit just in case
it breaks stuff.