Feature: add an "Import Layers…" action.
This allows multiple layers to be loaded into the current document from one
or more files that MyPaint knows how to load. It's a bit like the GIMP's
"Open As Layers" command, and uses the same keypress (Ctrl+Alt+O).

The imported layers all go into a single new group above the current layer.
Hopefully that's not too inconvenient, although it would be nice to have a
"Burst Group" command.