Dynamics: reset button clears the mapping graph.
When the brush dynamics are reset for a given input in the editor UI,
the graph was reset to some spurious zero entries. This was the cause
of some annoying knock-on effects in the brush engine, and since the
meaning of "reset" is kinda obvious, we should probably do the least
surprising thing.

Note that setting the dynamic's ±y slider to *exactly* zero does the
same graph reset (but see mypaint/mypaint#826 for a discussion of
whether that's a good idea or not)

Closes mypaint/mypaint#840.
Addresses mypaint/libmypaint#99 and mypaint/libmypaint#100. Sorta.