Debug menu: add "Vacuum Empty Tiles" command.
It would be useful to remove empty tiles every so often automatically,
so let's give our testers something to play with.

Not sure what the granularity of the background vacuums should be or
when they should happen yet. Backgrounded per layer, as a bunch of
idle commands? Before pushes onto the command stack?