lib.feedback: New mechanism for reporting feedback.
File loaders and savers will need to estimate the amount of work
they need to do, and count during the processing. It might be
useful to expand this interface later to report strings too,
but for now add a "simple" hierarchical report object which can
be passed around where we currently use feedback_cb arguments.

Part of the work for mypaint/mypaint#509.