Allow per-document settings.
This commit adds a settings dict to the current working document.  The dict
can be monitored for changes by any UI, and it gets serialized into .ora
files as a JSON file (data/mypaint-settings.json) that's indexed as a
MyPaint-specific extension to stack.xml.

These settings should be useful for palettes, perspective points, and other
goodies, provided the data you need to save is expressible in JSON.