Layers window: add thumbnail previews column.
This commit adds a slow, low-priority timer for regenerating layer preview
thumbnails as the layers are modified. Layers that are part of a document
tree now have thumbnails that update automatically, and these are shown in
the layers window if it's visible.

This may need a little testing, because it causes all the layer surface
mipmaps to regenerate → balloons memory. No worse than the zoomed-out
preview dockwindow though?