Bugfix: pass sized progress item to load_from_stream().
load_surface_from_pixbuf_file() was not putting a size estimate
on the progress item it was passed, causing an exception later
on. This was causing vector layer loads from .ora zipfiles
to be displayed with a "question mark box" fallback.
Possibly was breaking more.

Fix by adding a non-size assertion somewhere understandable,
and doing the sizing operation with a stat().