Widen layers list rows for future work (discuss?)
Some people might consider this commit inelegant or a bad use of sidebar
space. Best I write an explanation of what's going on here.

This commit makes each row of the layers list wider, but more informative.
I'm doing this to support future work.

(Also, double-clicking the name invokes the new layer properties dialog
now, but that's probably gonna be way less controversial.)

My goals with the widening are touch-friendliness and supporting a
potential Layer Sets feature. If I add that feature, the opacity slider
and the mode dropdown's functionality may move onto the layer properties
widget and be merely represented in the list as indicators or texts.

The click handlers for the flag and name columns may eventually fire the
new property widget in a popover. Popovers are pretty, and feel exactly
right for this context! However, popovers from sidebar widgets can't
happen while the workspace module still allows toolstacks in floating
windows ☹