EXPERIMENTAL: add bulk layer visibility switching.
This commit adds controls for switching between different views of the
document. Views describe which layers are visible together.

The list of views is saved to and restored from current .ORA file, but via
the document-specific settings dict. It should be considered an
experimental MyPaint-only feature for now.

Open questions:

1. Should views always be locked? This may make it simpler to use the
   feature correctly, at the cost of adding some potential confusion.

2. Is the UI placement in the layers dockable correct? Should it become
   button-and-dialog?

3. Do we need better icons?