Layers window: put mode and opacity on same line.
Put the mode dropdown and the opacity control onto the same line to
save space for the layer view switcher controls that are to come.

Mode dropdown is now 2-column, and uses smaller text.

Opacity control is now a GtkScaleButton with a dropdown scale. The
internal model for the opacity now uses a GtkAdjustment.

Update copyright.