Layers list: announce branch visibility/locked deeply.
Fudge a row-changed for all descendent layers of a group to make sure that
the eye or lock icons get updated immediately.

From GTK's point of view, this is smelly code given the TreeModel we use.
However, MyPaint treats the TreeView and its TreeModel as a big black box
labelled "View" much of the time. So it's OK. Honest.