Rework rendering: extract ops, tighter tile loop.
Update the rendering to add scope for some efficiency gains. The rendering
code may also be more understandable, more consistent, and less repetitious
now. YMMV, of course, but at least the mess is now a bit more centralized.

* Layers no longer implement TileBlittable or TileRenderable.
* Instead, they implement a new Renderable interface.
* Rendering is now implemented by methods of RootLayerStack.
* Rendered tiles can now stay in the render cache longer.
  - The cache is primarily keyed by tile index and mipmap level now.
  - Thus, drawing and painting now drops dirty cache tiles by position.
  - The cache is still cleared fully by layer property changes etc.
* Saving will have gained a small amount of efficiency.

Rendering now asks the layer tree up-front for a list of render ops to be
executed as a little program inside the tile loop.  The render loop is now
tighter since it longer has any recursive drilldowns executed for each tile
composite op.

The tighter this gets, the more feasible it becomes to do the tile loop in
C++ (possibly with a dropped GIL, even).