brushmanager: add doctests and a test brushpack.
MyPaint will needs to be able to load brushpacks saved under old Python
versions, so add some test data to prove that it will be able to.

There are also a ton of file and zipfile-specific unicode encoding issues
that the doctests will help fix.

With this commit, all the new doctests run in Py2, but not Py3.