filehandling: use filename_to_unicode (check me!)
Use our tested, compat version of g_filename_to_utf8() for interrogating
file dialogs. In Py3 it's a no-op as they mostly seem to be returning
Unicode anyway.

If this turns out to be a bad idea with some locales on Windows, it's in
its own commit so it can be reverted cleanly.