CODEQUAL: let flake8 run globally & generate useful output.
Now that the number of coding style violations is down to an approachable
level, it's time to automate the Python linter checks and get serious
about making the code consistent. I'm investigating Hound-CI for this atm.

Part of that is allowing "flake8" to run naturally over the entire project.
I've been using https://github.com/nvie/vim-flake8 for ages for new code,
but it's very piecemeal. It's also why you see all those "flake8" commits
in the recent history.

The code complexity target is arbitrary. It's chosen to match the current
MacCabe level of the layer stack's rendering method, and there is no magic
about it. Change it if it's too restrictive.

Anyway, with this commit, you can now run a recent flake8 at the top level
of your cloned repository without any special parameters, and get a list
of problems that can be usefully fixed.  And you should.  If you're
contributing a code-quality fix, please submit it as a PR in its own
right so that it can be reviewed and tested independently of other commits.

The goal of all of this is to create a codebase with an enforceable and
defined coding style, and no dead code or vars. I hope it'll reduce the
line churn in the long run too, to make fork maintenance easier.