fake pressure/rotation controller
Instead of sending static 0.5 for
devices that don't support it, we can
send anything we want, control it with
a slider and hotkeys.  This opens up a lot
of expression for users with mice or other
unsupported devices.

Also, this may be unpopular, but it is time
to free up modifiers shift, control, and alt.
Having these free allows us to re-use keys
to support on-the-fly brush adjustments like
changing colors, size, opacity, etc, while
mid-stroke, and use buttons that are
conveniently located