Add blend modes to flood fill
Add a singleton BlendModes instance
to the fill mode, allowing it to maintain
its own blend mode state and receive mode
change callbacks when active.

Update cursor icons to reflect mode state.

Add adjustable blend mode to floodfill call stack

The fill tool can now erase and lock the alpha of the
destination layer (the layer receiving the fill).

If eraser or lock alpha mode is active in conjunction
with the fill-to-new-layer option, a layer is created
but the fill is not actually performed (no point).

Make it so that non-permissible fills are actually
not performed (e.g. filling into a hidden layer).

closes #474