Prevent unnecessary tile creation when fill erasing
When erasing, there is no point in creating new tiles in the
destination surface (erasing an empty tile is not very useful).
The composite function now checks for tile existence in the
destination surface, and skips it fails.

Add tracking of the bounding box of affected tiles, instead
of the raw fill tiles, and only use the bounding box of the tiles
that are actually affected by the operation.

Remove redundant mipmap dirty marking - this is done in the tile
context manager regardless of use.