Make new C++ code style-consistent
There is no set convention or styleguide (as of writing)
for the C/C++ code in neither mypaint nor libmypaint.
This commit does not change that.

New files are updated with clang-format, using a configuration
which takes the most common occurences of brace-placement,
indentation, spacing etc. that can be found in current code.