Add option for creation/selection layer blinking
Layer blinking shows a layer in isolation for a short
time period when it is created or selected.

While useful for navigating a file with a ton of layers,
it can be annoying when wanting to switch between,
and work on, many layers quickly. This adds the option
to toggle this behavior in the preferences window.