Add cancellation and status info for fill
Add the ability to cancel a fill running for longer than
a certain time period. The fill runs in a separate thread
and the caller gets a handle that can be used to cancel
the operation.

New strings added for status messages about the progress
of the fill, shown in the cancellation/status dialog.

Add missing "enabled" check in keyboard manager.

Update tests