About window: fix double escaping
Some strings in the About dialog were escaped in the input
as well as in the gtk implementation, causing some unicode
to render as escaped html markup.

New comments indicate which functions escapes input before
it is used/displayed.

Thanks to Sabri Ünal (libreajans) for finding this problem
and what caused it.