setup.py: add new build flag for mypaintlib rpath
When building mypaintlib, the setup script uses pkg-config
to find the build information about the dependencies.

This adds a flag to add the paths of those dependencies
to the rpath of the library, making it easier to run
different installs against different versions of libmypaint.

Splits pkg-config data fetching into two steps to make sure
the libmypaint paths end up first in the library_dirs list.