Win32: set relative brush path at build time
Add an option to the BuildConfig setup command that takes a
static string and uses it instead of the absolute path fetched
via pkg-config.

Use new option to set a relative arch-dependent path
when building packages for the windows bundles.

Place the *-mypaint-git styrene dependency as primary again
(no functional difference in this case, as far as I know).