setup.py: fix locale files not being installed
When the build step was run separately from the
install step, the generated language catalogs were
not installed (i.e, when using --skip-build).

Refactors the BuildTranslation catalog generation
for easier external access to the paths used and
conditionally sets those paths in a new override
to install (if build_translation has not been run).

Also fixes an incorrect assertion when --dry-run
is enabled (uncaught if the build already exists)