setup.py: add option for disabling openmp
For some platforms (particularly musl-based), openmp is more
difficult to work with. This adds the --disable-openmp flag
to build_ext as a lazy workaround for building on such platforms.

Also allows the relevant flags to be set via environment variables,
to enable building using e.g. some older versions of clang.