Add syntax for installation-relative brushdir path
Brushdir paths leading with {installation-prefix} will be
prefixed with the installation path (prefix var) at runtime.

Right now, this is only useful for the AppImage build,
where this is more powerful than a static relative path
for debugging, as it permits running the build as either
the final AppImage, or from the extracted image contents,
without having to start the application from a particular
directory.