Remove the old SCons components and update docs
By now the transition to setuptools was finished a long time ago.
This removes the old build scripts, updates the debugging docs with
instructions for the current build system and removes outdated
(potentially confusing) references to scons.

Add a note in the build docs about accessing SCons files to uninstall