i18n: fix/rework fill option strings
Taking the opportunity to update these as they have
not been written to the po template file yet.

Made capitalization of option names consistent
with most other options in MyPaint.

Switched to the context "macro" for most strings,
made context info consistent and expanded descriptions
for labels/tooltips of the more complex options.

Replaced one outdated tooltip and reworked some tooltips
that were yet to be included in the translation files.