Alter build config to work with swig >= 4
A change in the generated module import makes builds
created with swig >= 4 fail to import mypaintlib.
This changes the installation layout to handle that.
The relevant swig change may have been introduced
in 3.0.9, but I have not verified this.

Note: In order to test-run from the source directory,
create a symlink to _mypaintlib.so in lib.
E.g (from repo root): ln -s ../_mypaintlib.so lib/_mypaintlib.so

For the above reason, lib/_mypaintlib.so is no longer removed
by scons - I think we can safely assume that anyone building
this branch will not have ancient build artifacts lying around.