i18n: add translation limitation option
Limit installation to configured locales.
All message objects are still built, but
only the ones added to lib.config in the
"build_config" build step are installed.

The build_config step now takes an option to filter out
locales to support, based on degree of completeness or
inclusion in a file listing locales to always include.