Fix translations when running from source tree
The old locale path does not work, as compiled
message objects are no longer placed in their
respective directory trees under the po directory,
but rather placed in a build-version-dependent
directory in the build output directory.

Create a symlink as part of the translation build
and use that when running from the source tree.