setup: revise static locales and share info cache
I forgot to include the file listing the always-include
exceptions for "build_config --translation-threshold=...".
Let's just use a list defined directly in setup.py instead.

Place cache file directly under the build directory so that
it can be used when building with both python2 and python3.

Also, run the threshold build option as part of the CI tests.