setup: refactor BuildConfig functions
Rewrite the replace function to run in a single
pass and use the 'with' statement for file ops.
It is now called 'instantiate_template' and
operates on a single input/output pair.

Switch substring replacement w. string formatting.

Add @staticmethod to a static method and add a
helper function for printing errors to stderr.