brusheditor: don't use +/- MAX_FLT in input limits
Before the brushlib split, unspecified limits were
indicated by None, because they were read directly
from the json spec. Now this is only indicated by
the limit values being +/- MAX_FLT.
Just assume that extremely high and low values
indicate that the limit is not specified.

This is the minimal solution to the wrong x range
being used i the editor; the change should arguably
be made directly in the BrushInputInfo class instead.