tree: pass render cache size in parameters
Since it is generally bad form to have lib code
be dependent on gui code, let the LRU capacity
parameter be passed via parameters instead of
being read from the preferences (in the lib code).

The default cache size is placed in a module constant
to remove the repetition, and a smaller cache size
(1/4 of default) is used for the scratchpad doc.