Improve performance of get_tiles_bbox, add tests
Calculating the bbox by getting the bounds of the
coordinates in a single pass improves performance
of this function by about ~5x.

The absolute difference is negligible when the number
of tiles is low, but is significant for larger amounts,
e.g. when using a line/curve tool w. symmetry enabled.