Add file compat check, warning dialog, cancellation
Issue warning and confirmation dialogs when attempting
to open OpenRaster files created in newer versions of
MyPaint. Cancelling the file opening will clear the
document.

There are two types of warnings: one for when only the
minor version of the file is larger than the minor version
of the software, and another (more severe) when the
major version of the file is larger than the major version
of the software. The warnings can be disabled independently.