Normalize hsv values in 2.0-alpha .ora strokemaps
Due to brush color now being consistently stored
without transformations, files created in the 2.x
alphas will have strokemap color mismatches.

This commit adds TEMPORARY support for these files,
inverting the color transform in the strokemaps on
load. Affected files are identified based on existence
and non-existence of xml attributes. It is assumed that
the standard eotf constant (2.2) was used for all of
the affected files (known exceptions: files created
in the first alpha will get skewed results - eotf was 2.4).

This commit should be REVERTED for the 2.1 or 2.2 release.
Users should be informed that the support is temporary and
be urged to re-save any affected files.