Base toolbar rotation and zooming on canvas center
There is seemingly no good way of having different behavior
without separate GtkAction's, so that is how this is now solved.

When accessed from the menus, rotation and zooming are still based
on the cursor location, as having a separate action there would mean
that any keyboard shortcuts for the regular actions wouldn't show up.