Optimize/Cull mkmamurk backgrounds
These backgrounds took up a huge amount of space compared to the rest
of the files shipped with MyPaint. These are better provided through
separate background pack downloads, since most of them are basically
duplicates with adjusted hue/brightness/contrast.

A representative selection is kept, and optimized to keep sizes
reasonably small. This reduces the size of these backgrounds from
26.2 MB to 1.7 MB (mostly because of the culling).