Refactor gui.workspace - split into more modules
To reduce the unwieldiness of the gui.workspace module (~2500 lines),
the ToolStack and ToolStackWindow classes, with associated helpers
and module constants, are split out into a new gui.toolstack module.
The positioning functions are moved to the gui.windowing module.

The use of a class-scoped GObject property for the workspace refs are
replaced with weakrefs (the plan may have been to expose the property
to be used in glade, but it appears this did not happen).