i18n: use square brackets in report template
The string isn't actually formatted with .format or .format_map,
so using braces for the example title has caused some confusion
in the translations. The braces are now replaced by square brackets,
both in the source string and the translation strings.

This is done directly, so the translations are not marked as fuzzy.
Reporters should be encouraged to use english when reporting issues
regardless, so leaving the original english string where it was not
translated is not really a big deal.