svg icons: fix xlink namespace not being specified
Export script adjusted and all icons re-exported.

librsvg 2.48.0 is stricter about namespaces being specified,
making this necessary for any of the icons that make use of
xlink:href (quite a few). Also includes some minor tweaks that
had not been exported to the actual icon files before.

Thanks to lazka for drawing attention to this.