Update icon extraction script
Remove invalid predicate for the rectangle deletion - any rectangle in
an icon, matching the icon's size, is assumed to be a bounding box.

Replace ``iteritems`` with ``items`` to make the script run w. Py3.
Never really made any sense to use iteritems for a dict w. <10 items.

Set more aggressive optimization options for scour (this level of
control may not have been available when the script was created).

Styles are cleaned up after the inkscape export step, making sure to
get rid of the unused attributes that are introduced after e.g.
stroke-to-path conversions.

Require py3 (>=3.5) to run, to make use of the improved handling of
stdin/stdout in the subprocess calls. This is not a build-time script;
expecting at least 3.5 to be available should not be controversial.