io: fix brush groups encoding bugs (Windows)
The encoding of 'order.conf' was not explicitly set to utf-8.
When the file was saved on Windows, the encoding would not generally
be utf-8, leading to crashes upon subsequent loads if any group or
brush name contained characters that could not be decoded w. utf-8.

This makes the encoding explicit on write, and handles the cases where
incorrect encoding was used previously (the latter should probably be
dropped in a few versions, as the bug was introduced in 2018 and was
only part of the 2.0.0 release).