i18n: fix access to libmypaint message catalogs
Answers the "open question" in the comment to the gettext setup:
the libmypaint messages do need to be bound, unless they are
installed in a location where the gettext build is configured
to look (such as /usr/share/locale/). That is why they have not
worked correctly in the appimages _unless_ the same messages
were already available on the system, in a standard path.

libmypaint's message domain is now always bound with bindtextdomain,
where the path is either configured at build time, or using the same
path as the one used for mypaint.

New options are added to build_config for this purpose:

--libmypaint-locale-path=
--libmypaint-locale-path-from-pkgconf

The options are mutually exclusive and if neither is supplied, the
path that is used for mypaint will also be used for libmypaint.

The version of libmypaint used is now also written to lib/config.py,
so that the correct domain can be bound in lib/gettext_setup.py.