Fix failing DocPaint tests
There were two issues with this test. One was that the reference
images were outdated, as they were created before the grey-pixel
artifacts were (mostly) fixed in libmypaint. The other was that
the new color sampling uses `rand()`, which means that the tests
will only be consistent if:

* rand() is thread-local, with a consistent starting seed, and
  with no other part of the program changing that seed.

* openmp is disabled, since different tile orders will yield
  (slightly) different results

Options are to either disable smudging or disable the new sampling
by explicitly setting `paint_mode` to 0. Here we choose the latter,
and update the reference images.