Update all HCY adjusters when gamut mask changes
This is a minimal change (code-wise) to fix the problem with
gamut masks only being updated for the dockable HCY adjuster
and not the quick adjuster.

In the long term the entire system needs to be refactored
to get rid of the cyclically dependent mixin classes, and
to remove hacky workarounds like this.