Ignore keyboard events in layer list
This is needed when widgets can receive keyboard events, since
the default list navigation is not handled, in terms of updating
the active layer to match the widget state. At some point, making
the layer state reflect the widget state directly should be
implemented instead, so this workaround can be removed.