Add slider with keyboard input (InputSlider)
Adds a new widget which is a single-child container, normally holding
a GtkScale instance (a slider), switching it out for a GtkSpinButton
when double clicking the slider - for numeric entry w. keyboard.

A bit of metaclass magic allows the new widget to act as a normal
scale with regards to instantiations and attribute access.
For that reason, another metaclass-related function from the "six"
library is added (under the MIT license), and the gibindings module
is modified to initialize all Gtk/Gdk type data, necessary for the
GObject.Property creation, in the new class, to work.