gui.mode: clean/fix `initial_modifiers` handling
For non spring-loaded modes, the handling of initial_modifiers was
not updated to take into account that the attribute is always defined
in the constructor (changed in 2014, two years after being added).
As a consequence, the "initial_modifiers != 0" condition would pass
when it shouldn't, resulting in the subsequent bit operation failing.

This commit also removes an unused local variable.