buttonmap: use mode's `get_name` as label fallback
The ACTION_NAME has to be defined in order for a mode class
to be registered, but when the value does not actually refer
to a GtkAction (usually defined in resources.xml) the fallback
label used to be the action name itself. It now uses the get_name
class method of the mode to get the fallback string instead.