Add canvas-drag brush resizing
New mode for changing the brush radius by click-dragging.
The radius is visualized by an overlay that can exceed the maximum
size of the regular brush cursors (especially when zoomed in).
The actual radius is updated at the end of the drag (for performance).

Precision mode can be enabled/disabled by toggling the state of a
modifier key (currently hardcoded to shift). If the mod key is part
of the combination bound to the resize mode, releasing the key will
enable precision mode and pressing it will disable it again.