Semi-revert previous last-event recording change
Commit of initial change: 521a828af761c90af43de9496543fe048c2f0112

Due to memory reclamation, storing even the wrapped events is
not safe (I should have tested this more thoroughly). Back to
unpacking on every call, but stores the values as a single tuple
in a regular attribute on the tileddrawwidget, rather than three
class-mangled attributes.

It remains that recording the event info on every motion event is
still pretty expensive, and should be improved one way or another.