gui.symmetry: replace center coord. input widgets
With keyboard input prioritizing input widgets, there is no longer
any need for the button+dialog workaround (housing the input widget
in modal dialogs). Instead, spin buttons are now used directly.

This makes it easier to set values directly, but also improves the
performance of moving the symmetry center, because the string
formatting for the button labels is no longer run.