gui.mode: add method `popped` to InteractionMode
Like `enter` and `leave`, this method can be overridden by mode
classes to perform specific tasks or cleanup when the mode has
been removed from the stack (but which should not be done in
`leave`, in case the mode has only been pushed down). It is always
called after `leave`, when a mode is popped from the stack.

Although modes can check if they still exist in the mode stack in
their `leave` implementations, this is cleaner overall.