gui.symmetry: refactor to improve gui performance
Most of the data is moved to the overlay instance, since it is always
needed there. The edit mode gets what little state info it needs from
the symmetry overlay instance of its TiledDrawWidget. The state data
is only recalculated when necessary.

Most importantly, the redraw code is changed so the entire canvas
is no longer invalidated on every symmetry state change when editing.
The invalidation rectangles can still be made better for long,
angled axis lines.