gui.symmetry: change line styles for axis lines
The drop shadow style, though nice-looking, can make it difficult
to see the symmetry axes when drawn on very light backdrops.
The drop shadows are also quite expensive to render.

This changes the style of the axis lines to alternating light and
dark dashes, making it discernable on dark, light and gray backdrops.