gui.mode: speed up motion_notify_cb
Local `_in_drag` boolean is now checked directly, to avoid the
overhead of the `in_drag` property. Though not significant for
most cases, for motion events it can add up quickly.

The setup of start_x/start_y attributes is refactored to ensure
that `_in_drag` implies those attributes being set up, so that
the conditional in the motion cb can be dropped.