gui.topbar: fix args in call to Gtk.Menu.popup
The call was missing either the arguments for time or for user data.
Call is now consistent with the other (except for time/button args).

The previous bindings may have been more lenient about missing
parameters, but that has not been the case since at least v1.2.0
going by issue #1099

Not sure if this codepath can even be triggered anymore, but might as
well fix it, in case opening the menu w. keys is made possible again.

Closes #1099