Refactor module & resource path setup/handling
Instead of making assumptions about the installation structure
in the startup script, the configuration used when the script
is installed is now used to set up paths relative to the install
script location for the resources and modules required. The brush
path is still handled separately.

The special handling for win32 builds is dropped - it is no longer
relevant since starting to use msys2 bundling for the Windows builds.