gui.windowing: remove UTILITY hint from SubWindow
The hint caused Gnome >= 3.30 (and possible older versions as well) to
spawn the icon editor window under the brush editor window when opened
from the latter (not from the menu). While this may be an upstream bug
in the Gnome WM, removing the hint does not appear to have any negative
impact when using other WM's, including in Windows 10.

Closes #1110