gui.picker: handle touch events in click handler
Fixes an assertion error when using touch devices to click the color
and context pickers on the footer bar. Not sure when/if the assertion
'the docs lie' was true, but the comment presumably refers to docs
that would be at least 5½ years old at this point - the documentation
for the current stable gtk3 does not mention anything about only
BUTTON_RELEASE events being emitted for the clicked signal handlers.

Closes #1123